+++ KoS.Apps.SharePoint.SmartCAML/changelog.md
@@ -2,3 +2,4 @@
 #### v. 1.8.0
 + New application icon and logo
 + Send update and install information to the server
++ Support 'Membership' operator